Fix #2754: make sensors_battery().percent return a float on all OSes

+++ psutil/_pslinux.py
@@ -1419,7 +1419,7 @@ def sensors_battery():
         except ZeroDivisionError:
             percent = 0.0
     else:
-        percent = int(cat(root + "/capacity", fallback=-1))
+        percent = float(cat(root + "/capacity", fallback=-1))
         if percent == -1:
             return None
